+++ src/main/java/dinkplugin/message/DiscordMessageHandler.java
@@ -20,17 +20,27 @@ import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.tuple.Pair;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.VisibleForTesting;
 
+import javax.imageio.ImageIO;
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import java.awt.Image;
+import java.awt.image.BufferedImage;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionException;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -95,24 +105,22 @@ public class DiscordMessageHandler {
 
         if (sendImage) {
             String screenshotFile = mBody.getType().getScreenshot();
-            drawManager.requestNextFrameListener(image -> {
-                try {
-                    byte[] imageBytes = Utils.convertImageToByteArray(ImageUtil.bufferedImageFromImage(image));
-
-                    reqBodyBuilder.addFormDataPart(
+            captureScreenshot(drawManager, config.screenshotScale() / 100.0)
+                .thenApply(image -> reqBodyBuilder
+                    .addFormDataPart(
                         "file",
                         screenshotFile,
                         RequestBody.create(
-                            MediaType.parse("image/png"),
-                            imageBytes
+                            MediaType.parse("image/" + image.getKey()),
+                            image.getValue()
                         )
-                    );
-                } catch (Exception e) {
+                    )
+                )
+                .exceptionally(e -> {
                     log.warn("There was an error creating bytes from captured image", e);
-                } finally {
-                    sendToMultiple(urlList, reqBodyBuilder);
-                }
-            });
+                    return reqBodyBuilder;
+                })
+                .thenAcceptAsync(body -> sendToMultiple(urlList, body));
         } else {
             sendToMultiple(urlList, reqBodyBuilder);
         }
@@ -160,6 +168,49 @@ public class DiscordMessageHandler {
                 }
             }
         });
+    }
+
+    /**
+     * Captures the next frame and applies the specified rescaling
+     * while abiding by {@link Embed#MAX_IMAGE_SIZE}.
+     *
+     * @param drawManager  {@link DrawManager}
+     * @param scalePercent {@link DinkPluginConfig#screenshotScale()} divided by 100.0
+     * @return future of the image byte array by the image format name
+     * @apiNote scalePercent should be in (0, 1]
+     * @implNote the image format is either "png" (lossless) or "jpeg" (lossy), both of which can be used in MIME type
+     */
+    private static CompletableFuture<Map.Entry<String, byte[]>> captureScreenshot(DrawManager drawManager, double scalePercent) {
+        CompletableFuture<Image> future = new CompletableFuture<>();
+        drawManager.requestNextFrameListener(future::complete);
+        return future.thenApply(ImageUtil::bufferedImageFromImage)
+            .thenApply(input -> Utils.rescale(input, scalePercent))
+            .thenApply(image -> {
+                try {
+                    String format = "png"; // lossless
+                    return Pair.of(format, Utils.convertImageToByteArray(image, format));
+                } catch (IOException e) {
+                    throw new CompletionException("Could not convert image to byte array", e);
+                }
+            })
+            .thenApply(pair -> {
+                byte[] bytes = pair.getValue();
+                int n = bytes.length;
+                if (n <= Embed.MAX_IMAGE_SIZE)
+                    return pair; // already compliant; no further rescale necessary
+
+                // calculate scale factor to comply with MAX_IMAGE_SIZE
+                double factor = Math.sqrt(1.0 * Embed.MAX_IMAGE_SIZE / n);
+
+                // bytes => original image => rescaled image => updated bytes
+                try (InputStream is = new ByteArrayInputStream(bytes)) {
+                    String format = "jpeg"; // lossy
+                    BufferedImage rescaled = Utils.rescale(ImageIO.read(is), factor);
+                    return Pair.of(format, Utils.convertImageToByteArray(rescaled, format));
+                } catch (Exception e) {
+                    throw new CompletionException("Failed to resize image below Discord size limit", e);
+                }
+            });
     }
 
     private static NotificationBody<?> injectContent(@NotNull NotificationBody<?> body, boolean screenshot, DinkPluginConfig config) {